move AutoChunkProcessor class to runtime-manager and reuse it
